requirements.txt
streamlit==1.28.1
huggingface-hub==0.19.4
python-dotenv==1.0.0
gradio==4.44.1
requests==2.32.3
pydantic==2.8.2